+++ src/components/common/chord-diagram/ChordDiagram.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useRef } from 'react';
-import { drawBarre, drawFretNumber, drawFrets, drawMuteStringMark, drawOpenStringMark, drawPosition, drawStrings } from './utils';
+import { drawDiagram } from './frawing-main';
 
 interface Props {
 
@@ -11,25 +11,6 @@ export const ChordDiagram = ({ }: Props) => {
 
     const canvasRef = useRef<HTMLCanvasElement>(null);
     const containerRef = useRef<HTMLDivElement>(null);
-
-    const drawDiagram = (ctx: CanvasRenderingContext2D, width: number, height: number) => {
-        ctx.clearRect(0, 0, width, height);
-        const canvasSize = { w: width, h: height };
-
-        drawStrings(ctx, canvasSize);
-        drawFrets(ctx, canvasSize, false);
-
-        drawBarre(ctx, canvasSize, 1, 1, 5);
-
-        drawPosition(ctx, canvasSize, 2, 3);
-        drawPosition(ctx, canvasSize, 3, 2);
-        drawPosition(ctx, canvasSize, 4, 3);
-
-        drawOpenStringMark(ctx, canvasSize, 0);
-        drawMuteStringMark(ctx, canvasSize, 5);
-
-        drawFretNumber(ctx, canvasSize, 5);
-    };
 
     useEffect(() => {
         const canvas = canvasRef.current;
