+++ src/components/common/chord-diagram/constants.ts
@@ -1,0 +1,32 @@
+/** 指板の左端の位置 */
+export const START_X = 40;
+/** 指板の上端の位置 */
+export const START_Y = 40;
+/** 指板の右端の位置 */
+export const END_X = 20;
+/** 指板の下端の位置 */
+export const END_Y = 50;
+
+/** 線の太さ */
+export const LINE_WIDTH = 2;
+
+/** 弦の数 */
+export const STRING_NUM = 6;
+/** 弦を右に余らせる長さ */
+export const STRING_ESCESS_WIDTH = 10;
+
+/** フレットの数 */
+export const FRET_NUM = 5;
+
+/** 左側マーク(開放弦・ミュート)の指板左端からの距離 */
+export const LEFT_MARK_INTERVAL = 6;
+/** 左側マーク(開放弦・ミュート)のサイズ */
+export const LEFT_MARK_SIZE = 10;
+
+/** ポジションマークのサイズのフレット幅に対する割合 */
+export const POSITION_MARK_RATIO = 0.42;
+
+/** フレット番号の指板左端からの距離 */
+export const FRET_NUMBER_INTERVAL = 16;
+/** フレット番号のテキストサイズ(px) */
+export const FRET_NUMBER_SIZE = 18;
